small changes over Sultan code
CIC
user_functions
