+++ auxiliary.c
@@ -753,7 +753,8 @@ void get_collapsed_mass_boxb(float* halo_box,Halo_t *halo, long int nhalos){
 }    
 
 
-void CIC_smoothing(float x1, float y1, float z1, float map_in, float *map_out){
+/* distributes a "cell mass" that is off center between the surrounding cells... */
+void CIC_smoothing(float x1, float y1, float z1, float map_in, float *map_out, long int N){
 
   long int i1,j1,p1,x,y,z;
   x = (long int) x1;
@@ -782,15 +783,15 @@ void CIC_smoothing(float x1, float y1, float z1, float map_in, float *map_out){
           z1 = fabsf(z1);
         }
         /*********************check periodic boundaries*************************/
-        x=check_borders(x,global_N_halo);
-        y=check_borders(y,global_N_halo);
-        z=check_borders(z,global_N_halo);
-        /****************chaning ratios for different directions ***************/
+        x=check_borders(x,N);
+        y=check_borders(y,N);
+        z=check_borders(z,N);
+        /****************changing ratios for different directions ***************/
         if(i1 == 1) x1 = 1.0 - x1;
         if(j1 == 1) y1 = 1.0 - y1;
         if(p1 == 1) z1 = 1.0 - z1;
-        map_out[x*global_N_halo*global_N_halo+y*global_N_halo+z] += (1.0 - x1)*(1.0 - y1)*(1.0 - z1)*map_in;
-        /********************* go back the parent cell original coordinates and ratios ***************************/
+        map_out[x*N*N+y*N+z] += (1.0 - x1)*(1.0 - y1)*(1.0 - z1)*map_in;
+        /********************* go back to the parent cell original coordinates and ratios ***************************/
         x -= i1;
         y -= j1;
         z -= p1;
