+++ simfast21.ini
@@ -50,7 +50,7 @@ STb = 0.34
 STc = 0.81
 # Use subgrid to account for unresolved halos?
 Use_subgrid = T
-# Maximum halo size in Mpc
+# Maximum halo "size" in Mpc (maximum size for sphere)
 halo_Rmax = 40
 # Minimum halo radius in cell size units above which we can use the halo excursion set formalism (e.g. sets the minimum resolution for the method)
 halo_Rmin_dx = 2.0
